Refactor : 내 일정 utils 분리
